handleDragOver refactor
